Renamed package
I'm trying to create a mathewsmobile "brand" so I'm moving anything that will be released to have an updated package. I bought the mathewsmobile.com domain, just haven't done anything with it yet.

Note: also updated the google-services.json file for Firebase after adding this as a new project in the console.
