--- app/src/main/java/com/boboddy/pwned/api/PwnedApi.kt
@@ -1,6 +1,0 @@
-package com.boboddy.pwned.api
-
-/**
- * Created by nicke on 2/4/2018.
- */
-
