--- app/src/main/java/com/boboddy/pwned/model/DataClass.kt
@@ -1,7 +1,0 @@
-package com.boboddy.pwned.model
-
-/**
- * Created by nicke on 2/4/2018.
- */
-
-class DataClass(val type: String)
