+++ app/src/main/java/com/mathewsmobile/pwned/model/BreachesResponse.kt
@@ -1,0 +1,7 @@
+package com.mathewsmobile.pwned.model
+
+/**
+ * Created by nicke on 2/4/2018.
+ */
+
+class BreachesResponse(val breaches: List<Breach>)
