+++ app/src/main/res/layout/activity_info.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.boboddy.pwned.activities.InfoActivity">
+    tools:context="com.mathewsmobile.pwned.activities.InfoActivity">
 
     <ImageView
         android:id="@+id/info_logo"
