+++ app/src/main/res/layout/activity_pwn_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.boboddy.pwned.activities.PwnDetailActivity">
+    tools:context="com.mathewsmobile.pwned.activities.PwnDetailActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
