+++ app/src/main/res/layout/activity_pwned.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.boboddy.pwned.activities.PwnedActivity">
+    tools:context="com.mathewsmobile.pwned.activities.PwnedActivity">
 
     <TextView
         android:id="@+id/pwned_result"
