+++ app/src/main/res/layout/content_pwn_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.boboddy.pwned.activities.PwnDetailActivity"
+    tools:context="com.mathewsmobile.pwned.activities.PwnDetailActivity"
     tools:showIn="@layout/activity_pwn_detail">
 
     <TextView
